Add Ability to Toggle Power State of Lights
Add the --on & --off flags to the `set-` commands in the CLI client.
These include a workaround for strange bridge behavior.

Normally, when sending an --off flag with a transition-time, the light
will dim to a brightness of 1 and then turn off. Then when you turn the
light on, it will stay at a brightness of 1 instead of returning to it's
previous brightness.

We circumvent this by storing the off brightness & restoring it when the
light is turned back on(unless a brightness is specified).