Add CLI Command to Sync Redshift to Lights
Add a `redshift` command to the CLI client. This mode continuously
pulls the average color temperature from the daemon & uses the `-P` &
`-O TEMP` redshift flags to make your monitors match the color
temperature of your lights.

An `--interval` flag is available to change the time in between syncs,
although the cached BridgeState is used for the calculations so the
update times have a lower limit. Eventually more frequent syncs of the
light states will be made to the cache so even lower sync intervals will
be possible.