Cache Color Temperatures & Use for Average Calculation
Add a Map from Light IDs to the last set color temperature. When
calculating the average color temperature, we fall back to these cached
values if there are no powered-on lights in color temperature mode.

This makes the color temperature of several computers match after
switching to an RGB color and then starting `hkhue redshift` on the
other machine.