Add a SetGroupState Message for Clients
Add a new SetGroupState ClientMsg that tells the Daemon to set the light
state for every light in a group. The Daemon will either send a batched
update to the group, or update each light individually if different
state changes are needed for proper handling of the On/Off brightness
via the handlePowerBrightness function.