Add Group Syncing to the Faster Sync Thread
Change the lightsSyncThread to a partialSyncThread that also hits the
groups endpoint to update the bridge groups. These endpoints are less
resource-intense for the Bridge than the full bridge sync endpoint.

Change the getGroups function into a getFullGroupStates function that
returns a parsed BridgeGroup Map instead of a generic JSON Value.